// src/app/model/polineuropatia.model.ts
export class Polineuropatia {

  constructor(Edad: number, Genero: number, flgCovid: number, UCI: number, flgDiabetes: number, TipoExtremidad: number, Latencia_1_der: number, Amplitud_1_der: number, Latencia_2_der: number, Amplitud_2_der: number, Latencia_1_izq: number, Amplitud_1_izq: number, Latencia_2_izq: number, Amplitud_2_izq: number, Amp_Needle_1_der: number, Amp_Needle_2_der: number, Amp_Needle_3_der: number, Amp_Needle_1_izq: number, Amp_Needle_2_izq: number, Amp_Needle_3_izq: number) {
    this.Edad = Edad;
    this.Genero = Genero;
    this.flgCovid = flgCovid;
    this.UCI = UCI;
    this.flgDiabetes = flgDiabetes;
    this.TipoExtremidad = TipoExtremidad;
    this.Latencia_1_der = Latencia_1_der;
    this.Amplitud_1_der = Amplitud_1_der;
    this.Latencia_2_der = Latencia_2_der;
    this.Amplitud_2_der = Amplitud_2_der;
    this.Latencia_1_izq = Latencia_1_izq;
    this.Amplitud_1_izq = Amplitud_1_izq;
    this.Latencia_2_izq = Latencia_2_izq;
    this.Amplitud_2_izq = Amplitud_2_izq;
    this.Amp_Needle_1_der = Amp_Needle_1_der;
    this.Amp_Needle_2_der = Amp_Needle_2_der;
    this.Amp_Needle_3_der = Amp_Needle_3_der;
    this.Amp_Needle_1_izq = Amp_Needle_1_izq;
    this.Amp_Needle_2_izq = Amp_Needle_2_izq;
    this.Amp_Needle_3_izq = Amp_Needle_3_izq;
  }

  Edad: number;
  Genero: number;
  flgCovid: number;
  UCI: number;
  flgDiabetes: number;
  TipoExtremidad: number;
  Latencia_1_der: number;
  Amplitud_1_der: number;
  Latencia_2_der: number;
  Amplitud_2_der: number;
  Latencia_1_izq: number;
  Amplitud_1_izq: number;
  Latencia_2_izq: number;
  Amplitud_2_izq: number;
  Amp_Needle_1_der: number;
  Amp_Needle_2_der: number;
  Amp_Needle_3_der: number;
  Amp_Needle_1_izq: number;
  Amp_Needle_2_izq: number;
  Amp_Needle_3_izq: number;
}